Contrôle connexion dans SortieController.php

+++ src/Controller/SortieController.php
@@ -20,11 +20,6 @@ use Symfony\Component\Serializer\SerializerInterface;
 #[Route('/sortie')]
 class SortieController extends AbstractController
 {
-    private const MESSAGE_CREATION = 'La sortie a été créée';
-    private const MESSAGE_PUBLICATION = 'La sortie a été publiée';
-    private const MESSAGE_MODIFICATION = 'La sortie a été modifiée';
-    private const MESSAGE_SUPPRESSION = 'La sortie a été supprimée';
-
     private ControleSortie $service;
     public function __construct(ControleSortie $service)
     {
@@ -43,6 +38,10 @@ class SortieController extends AbstractController
                         ParticipantRepository $participantRepository, CampusRepository $campusRepository,
                         LieuRepository $lieuRepository, EtatRepository $etatRepository): Response
     {
+        if (!$this->getUser()) {
+            $this->addFlash('warning', $this->service::MESSAGE_LOGIN);
+            return $this->redirectToRoute('app_login', [], Response::HTTP_SEE_OTHER);
+        }
         $sortie = new Sortie();
         $form = $this->createForm(SortieType::class, $sortie);
         $form->handleRequest($request);
@@ -60,10 +59,10 @@ class SortieController extends AbstractController
 
             if ($form->get('enregistrer')->isClicked()) {
                 $sortie->setEtat($etatRepository->findOneBy(array('libelle' => 'Créée')));
-                $message = self::MESSAGE_CREATION;
+                $message = $this->service::MESSAGE_CREATION;
             } else if ($form->get('publier')->isClicked()) {
                 $sortie->setEtat($etatRepository->findOneBy(array('libelle' => 'Ouverte')));
-                $message = self::MESSAGE_PUBLICATION;
+                $message = $this->service::MESSAGE_PUBLICATION;
             }
             $sortieRepository->save($sortie, true);
 
@@ -80,7 +79,12 @@ class SortieController extends AbstractController
     #[Route('/{id}', name: 'app_sortie_show', requirements: ['id'=> '\d+'], methods: ['GET'])]
     public function show(Sortie $sortie): Response
     {
+        if (!$this->getUser()) {
+            $this->addFlash('warning', $this->service::MESSAGE_LOGIN);
+            return $this->redirectToRoute('app_login', [], Response::HTTP_SEE_OTHER);
+        }
         if (!$this->service->estAffichable($sortie)) {
+            $this->addFlash('notice', $this->service::MESSAGE_NON_AFFICHABLE);
             return $this->redirectToRoute('app_sortie_list', [], Response::HTTP_SEE_OTHER);
         }
         return $this->render('sortie/show.html.twig', [
@@ -134,14 +138,18 @@ class SortieController extends AbstractController
     public function edit(Request $request, Sortie $sortie, SortieRepository $sortieRepository,
                          LieuRepository $lieuRepository, EtatRepository $etatRepository): Response
     {
+        if (!$this->getUser()) {
+            $this->addFlash('warning', $this->service::MESSAGE_LOGIN);
+            return $this->redirectToRoute('app_login', [], Response::HTTP_SEE_OTHER);
+        }
         $form = $this->createForm(SortieType::class, $sortie);
         $form->handleRequest($request);
         $lieus = $sortie->getLieu()->getVille()->getLieus();
 
-        //Test sur l'id de l'utilisateur à voir
-        if ($sortie->getOrganisateur()->getId() === 0 ||
+        if ($this->getUser() === $sortie->getOrganisateur()->getId() ||
             !$this->service->estModifiable($sortie))
         {
+            $this->addFlash('warning', $this->service::MESSAGE_NON_MODIFIABLE);
             return $this->redirectToRoute('app_sortie_list', [], Response::HTTP_SEE_OTHER);
         }
 
@@ -150,11 +158,11 @@ class SortieController extends AbstractController
             $lieu = $lieuRepository->find((int)$request->request->get('sortie')['lieu']);
             $sortie->setLieu($lieu);
             if ($form->get('enregistrer')->isClicked()) {
-                $message = self::MESSAGE_MODIFICATION;
+                $message = $this->service::MESSAGE_MODIFICATION;
             }
             else if ($form->get('publier')->isClicked()) {
                 $sortie->setEtat($etatRepository->findOneBy(array('libelle' => 'Ouverte')));
-                $message = self::MESSAGE_PUBLICATION;
+                $message = $this->service::MESSAGE_PUBLICATION;
             }
 
             $sortieRepository->save($sortie, true);
@@ -172,11 +180,15 @@ class SortieController extends AbstractController
     #[Route('/{id}', name: 'app_sortie_delete', methods: ['POST'])]
     public function delete(Request $request, Sortie $sortie, SortieRepository $sortieRepository): Response
     {
+        if (!$this->getUser()) {
+            $this->addFlash('warning', $this->service::MESSAGE_LOGIN);
+            return $this->redirectToRoute('app_login', [], Response::HTTP_SEE_OTHER);
+        }
         if ($this->isCsrfTokenValid('delete'.$sortie->getId(), $request->request->get('_token')) &&
             $this->service->estModifiable($sortie))
         {
             $sortieRepository->remove($sortie, true);
-            $this->addFlash('notice', self::MESSAGE_SUPPRESSION);
+            $this->addFlash('notice', $this->service::MESSAGE_SUPPRESSION);
         }
 
         return $this->redirectToRoute('app_sortie_list', [], Response::HTTP_SEE_OTHER);
